+++ src/api/hook/router/UserAPI.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useContext } from 'react';
 import axios from 'axios';
 import { UserContext } from './UserContext';
-import Utils from './Utils';
+import Utils from '@/service/Utils';
 
 const UserAPI = () => {
   const { setUserData, setError, setAllUserData } = useContext(UserContext);
