+++ src/components/Calendar/EventDetailTitle.tsx
@@ -6,8 +6,8 @@ import Modal from 'react-modal';
 import EditDelModal from '@/components/Modal/EditDelModal';
 import IndexButton from '@/components/Header/IndexButton';
 import LeftButton from '@/components/Header/LeftButton';
-import UserAPI from '@/service/UserAPI';
-import { UserContext } from '@/service/UserContext';
+import UserAPI from '@/api/hook/router/UserAPI';
+import { UserContext } from '@/api/hook/router/UserContext';
 import * as S from '@/styles/calendar/EventDetailTitle.styled';
 import { adminModalStyles } from '@/styles/calendar/EventDetailTitle.styled';
 import React, { useContext, useState } from 'react';
