+++ src/pages/Member.tsx
@@ -5,8 +5,8 @@ import theme from '@/styles/theme';
 import MemberHeader from '@/components/Member/MemberHeader';
 import Category from '@/components/Member/Category';
 import MemberName from '@/components/Member/MemberName';
-import UserAPI from '@/service/UserAPI';
-import { UserContext } from '@/service/UserContext';
+import UserAPI from '@/api/hook/router/UserAPI';
+import { UserContext } from '@/api/hook/router/UserContext';
 import useCustomBack from '@/router/useCustomBack';
 
 interface User {
