+++ src/pages/MyPage.tsx
@@ -17,8 +17,8 @@ import icPosition from '@/assets/images/ic_position.svg';
 import icEmail from '@/assets/images/ic_mail.svg';
 import icEdit from '@/assets/images/ic_edit.svg';
 import icLogout from '@/assets/images/ic_logout_white.svg';
-import { UserContext } from '@/service/UserContext';
-import UserAPI from '@/service/UserAPI';
+import { UserContext } from '@/api/hook/router/UserContext';
+import UserAPI from '@/api/hook/router/UserAPI';
 import handleLogout from '@/utils/handleLogout';
 import useCustomBack from '@/router/useCustomBack';
 import theme from '@/styles/theme';
