+++ src/pages/StudyDetail.jsx
@@ -3,8 +3,8 @@ import React, { useState, useEffect, useContext } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
 import { useNavigate, useParams } from 'react-router-dom';
-import UserAPI from '../service/UserAPI';
-import { UserContext } from '../service/UserContext';
+import UserAPI from '@/api/hook/router/UserAPI';
+import { UserContext } from '@/api/hook/router/UserContext';
 import NoticeHeader from '../components/Board/NoticeHeader';
 import AttachButton from '../components/Board/AttachButton';
 import CommentList from '../components/Board/CommentList';
